add connection menu into flow view
