+++ app/javascript/src/controller_services.js
@@ -25,6 +25,8 @@ const ActivatedMenu = require('./component_activated_menu');
 const DialogApiRequest = require('./component_dialog_api_request');
 const DefaultController = require('./controller_default');
 const ConnectorPath = require('./component_flow_connector_path');
+const PageMenu = require('./component_page_menu');
+const ConnectionMenu = require('./component_connection_menu');
 
 const COLUMN_SPACING = 100;
 const SELECTOR_FLOW_BRANCH = ".flow-branch";
@@ -57,8 +59,8 @@ ServicesController.edit = function() {
   view.$flowDetached = $("#flow-detached");
 
   createPageAdditionDialog(view);
-  createPageAdditionMenu(view);
-  createFlowItemMenus(view);
+  createPageMenus(view);
+  createConnectionMenus(view); 
 
   if(view.$flowOverview.length) {
     layoutFormFlowOverview(view);
@@ -117,170 +119,6 @@ class FlowConditionItem {
   }
 }
 
-
-/* VIEW SPECIFIC COMPONENT:
- * ------------------------
- * Control form step (add/edit/delete/preview...) menus
- **/
-class FlowItemMenu extends ActivatedMenu {
-  constructor($node, config) {
-    super($node, mergeObjects({
-      activator_classname: $node.data("activator-classname"),
-      container_id: $node.data("activated-menu-container-id"),
-      activator_text: $node.data("activator-text")
-    }, config));
-
-    $node.on("menuselect", (event, ui) => {
-      this.selection(event, ui.item);
-    });
-
-    this.activator.$node.addClass("FlowItemMenuActivator");
-    this.container.$node.addClass("FlowItemMenu");
-    this.uuid = $node.data("uuid");
-    this.title = $node.data("title");
-  }
-
-   // Handle item selections on the form step context menu elements.
-  selection(event, item) {
-    var action = item.data("action");
-
-    event.preventDefault();
-    switch(action) {
-      case "preview":
-           this.previewPage(item);
-           break;
-
-      case "add":
-           this.addPage(item);
-           break;
-
-      case "destination":
-           this.changeDestination(item);
-           break;
-
-      case "delete":
-           this.deleteItem(item);
-           break;
-
-      case "delete-api":
-           this.deleteItemApi(item);
-           break;
-
-      default: this.link(item);
-    }
-  }
-
-  link(element) {
-    var $link = element.find("> a");
-    location.href = $link.attr("href");
-  }
-
-  previewPage(element) {
-    var $link = element.find("> a");
-    window.open($link.attr("href"));
-  }
-
-  // Open the views Page Addition Menu
-  addPage(element) {
-    var menu = this._config.view.pageAdditionMenu;
-    menu.addPageAfter = this.uuid;
-    menu.open({
-      my: "left top",
-      at: "right top",
-      of: element
-    });
-  }
-
-  // Open an API request dialog to change destination
-  changeDestination(element) {
-    var view = this._config.view;
-    var $link = element.find("> a");
-    new DialogApiRequest($link.attr("href"), {
-      activator: $link,
-      buttons: [{
-        text: view.text.dialogs.button_change_destination,
-        click: function(dialog) {
-          dialog.$node.find("form").submit();
-        }
-      }, {
-        text: view.text.dialogs.button_cancel
-      }]
-    });
-  }
-
-  // Use standard delete modal to remove
-  deleteItem(element) {
-    var view = this._config.view;
-    var $link = element.find("> a");
-    view.dialogConfirmationDelete.open({
-      heading: view.text.dialogs.heading_delete.replace(/%{label}/, this.title),
-      ok: view.text.dialogs.button_delete_page
-      }, function() {
-        post($link.attr("href"), { _method: "delete" });
-    });
-  }
-
-  deleteItemApi(element) {
-    var $link = element.find("> a");
-    new DialogApiRequest($link.attr("href"), {
-      activator: $link,
-      closeOnClickSelector: ".govuk-button",
-      build: function(dialog) {
-        // Find and correct (make work!) any method:delete links
-        dialog.$node.find("[data-method=delete]").on("click", function(e) {
-          e.preventDefault();
-          utilities.post(this.href, { _method: "delete" });
-        });
-      }
-    });
-  }
-}
-
-/* VIEW SPECIFIC COMPONENT:
- * ------------------------
- * Controls form step Add page functionality
- **/
-class PageAdditionMenu extends ActivatedMenu {
-  constructor($node, config) {
-    super($node, mergeObjects({
-      activator_classname: $node.data("activator-classname"),
-      container_id: $node.data("activated-menu-container-id"),
-      activator_text: $node.data("activator-text")
-    }, config));
-
-    this.container.$node.addClass("PageAdditionMenu");
-
-    // Register event handler for selection of menu item.
-    $node.on("menuselect", (event, ui) => {
-      this.selection(event, ui.item);
-    });
-  }
-
-  set addPageAfter(uuid) {
-    this._uuid = uuid;
-  }
-
-  get addPageAfter() {
-    return this._uuid;
-  }
-
-  selection(event, item) {
-    var dialog = this._config.view.pageAdditionDialog;
-    var $form = dialog.$form;
-
-    // Set the 'add_page_here' value to mark point of new page inclusion.
-    // Should be a uuid of previous page or blank if at end of form.
-    utilities.updateHiddenInputOnForm($form, "page[add_page_after]", this.addPageAfter);
-
-    // Then add any required values.
-    utilities.updateHiddenInputOnForm($form, "page[page_type]", item.data("page-type"));
-    utilities.updateHiddenInputOnForm($form, "page[component_type]", item.data("component-type"));
-
-    this._config.view.pageAdditionDialog.open();
-  }
-}
-
-
 /* VIEW SETUP FUNCTION:
  * --------------------
  * Finds the (in page) form that can add a new page and enhances with Dialog component
@@ -310,26 +148,43 @@ function createPageAdditionDialog(view) {
  * --------------------
  * Create the menu effect and required functionality for controlling and selecting new page types.
  **/
-function createPageAdditionMenu(view) {
-  var $menu = $("[data-component='PageAdditionMenu']"); // Expect only one
+// function createPageAdditionMenu(view) {
+//   var $menu = $("[data-component='PageAdditionMenu']"); // Expect only one
 
-  view.pageAdditionMenu = new PageAdditionMenu($menu, {
-    view: view,
-    selection_event: "PageAdditionMenuSelection",
-    menu: {
-      position: { at: "right+2 top-2" } // Position second-level menu in relation to first.
-    }
-  });
-}
+//   view.pageAdditionMenu = new PageAdditionMenu($menu, {
+//     view: view,
+//     selection_event: "PageAdditionMenuSelection",
+//     menu: {
+//       position: { at: "right+2 top-2" } // Position second-level menu in relation to first.
+//     }
+//   });
+// }
 
 
 /* VIEW SETUP FUNCTION:
  * --------------------
  * Create the context menus for each flow item within an overview layout.
  **/
-function createFlowItemMenus(view) {
+function createPageMenus(view) {
   $("[data-component='ItemActionMenu']").each((i, el) => {
-    var menu = new FlowItemMenu($(el), {
+    var menu = new PageMenu($(el), {
+      view: view,
+      preventDefault: true, // Stops the default action of triggering element.
+      menu: {
+        position: { at: "right+2 top-2" }
+      }
+    });
+
+    view.addLastPointHandler(menu.activator.$node);
+  });
+}
+/* VIEW SETUP FUNCTION:
+ * --------------------
+ * Create the connection menus for each flow item within an overview layout.
+ **/
+function createConnectionMenus(view) {
+  $("[data-component='ConnectionMenu']").each((i, el) => {
+    var menu = new ConnectionMenu($(el), {
       view: view,
       preventDefault: true, // Stops the default action of triggering element.
       menu: {
