+++ app/javascript/styles/_index.scss
@@ -714,7 +714,6 @@ html {
 
 .flow-page {
   height: 250px;
-  overflow: hidden;
   vertical-align: middle;
   width: 200px;
 }
