+++ app/views/services/_connection_menu.html.erb
@@ -1,0 +1,62 @@
+<ul class="govuk-navigation component-activated-menu"
+    data-activator-text="<%= t('pages.create') %>"
+    data-activator-classname="<%= item[:type] == 'flow.branch' ? 'branch-' : '' -%>connection-menu-activator"
+    data-component="ConnectionMenu"
+    data-title="<%= item[:title] %>"
+    data-uuid="<%= item[:uuid] %>">
+
+  <li data-action="none">
+    <span>Single question page</span>
+    <ul class="govuk-navigation">
+      <li data-page-type="singlequestion"
+          data-component-type="text"><a href="#add-page">Text</a></li>
+      <li data-page-type="singlequestion"
+          data-component-type="textarea"><a href="#add-page">Textarea</a></li>
+      <li data-page-type="singlequestion"
+          data-component-type="number"><a href="#add-page">Number</a></li>
+      <li data-page-type="singlequestion"
+          data-component-type="date"><a href="#add-page">Date</a></li>
+      <li data-page-type="singlequestion"
+          data-component-type="radios"><a href="#add-page">Radio buttons</a></li>
+      <li data-page-type="singlequestion"
+          data-component-type="checkboxes"><a href="#add-page">Checkboxes</a></li>
+      <li data-page-type="singlequestion"
+          data-component-type="upload"><a href="#add-page">File upload</a></li>
+    </ul>
+  </li>
+  
+  <li data-page-type="multiplequestions"><a href="#add-page">Multiple question page</a></li>
+  
+  <li data-page-type="content"><a href="#add-page">Content page</a></li>
+  <% if ENV['BRANCHING'] == 'enabled' %>
+
+    <li data-page-type="exit"><a href="#add-page">Exit page</a></li>
+
+    <% unless service.checkanswers_page.present? %>
+      <li data-page-type="checkanswers"><a href="#add-page">Check answers page</a></li>
+    <% end %>
+      
+    <% unless service.confirmation_page.present? %>
+      <li data-page-type="confirmation"><a href="#add-page">Confirmation page</a></li>
+    <% end %>
+
+    <% unless (item[:type] =~ /page.(checkanswers|confirmation|exit)/) || item[:type] == 'flow.branch' %>
+      <li data-action="link">
+        <%= link_to t('services.branch'),
+        new_branches_path(service.service_id, item[:uuid]) %>
+      </li>
+
+      <li data-action="destination">
+        <%= link_to t('actions.change_destination'),
+        new_api_service_flow_destination_path(service.service_id, item[:uuid]) %>
+      </li>
+    <% end %>
+
+  <% else %>
+    <li data-page-type="checkanswers"><a href="#add-page">Check answers page</a></li>
+    <li data-page-type="confirmation"><a href="#add-page">Confirmation page</a></li>
+  <% end %>
+
+
+
+</ul>
