+++ app/views/services/_flow_layout.html.erb
@@ -22,7 +22,7 @@
   <% else %>
 
     <% if item[:type] == 'flow.branch' %>
-      <article class="flow-item flow-branch" aria-label="TODO-need-editable-label" data-fb-id="<%= item[:uuid] %>">
+      <article class="flow-item flow-branch" aria-label="<%= item[:title] %>" data-fb-id="<%= item[:uuid] %>">
         <%= image_pack_tag "diamond.svg",  :usemap => "#thumbnail-#{index}" %>
         <map name="thumbnail-<%= index %>" aria-hidden="true">
           <area shape="poly" coords="0,72.5 100,0 200,72.5 100,145" alt="" href="<%= edit_branch_path(service.service_id, item[:uuid]) %>">
@@ -58,16 +58,19 @@
 
       <% if item[:next].empty? %>
         <%# Likely to be items like the end of a path (e.g. Confirmation page) %>
-        <article class="flow-item flow-page" aria-label="TODO-need-editable-label" data-fb-id="<%= item[:uuid] %>">
+        <article class="flow-item flow-page" aria-label="<%= item[:title] %>" data-fb-id="<%= item[:uuid] %>">
           <%= flow_thumbnail_link(item) %>
           <%= flow_text_link(item) %>
           <%= render partial: 'services/flow_page_menu', locals: { item: item } %>
         </article>
       <% else %>
-        <article class="flow-item flow-page" aria-label="TODO-need-editable-label" data-fb-id="<%= item[:uuid] %>" data-next="<%= item[:next] %>">
+        <article class="flow-item flow-page" aria-label="<%= item[:title] %>" data-fb-id="<%= item[:uuid] %>" data-next="<%= item[:next] %>">
           <%= flow_thumbnail_link(item) %>
           <%= flow_text_link(item) %>
           <%= render partial: 'services/flow_page_menu', locals: { item: item } %>
+          <% unless (item[:type] =~ /page.(checkanswers|confirmation|exit)/) %>
+            <%= render partial: 'services/connection_menu', locals: { item: item } %>
+          <% end %>
         </article>
       <% end %>
     <% end %>
