+++ app/views/services/edit.html.erb
@@ -40,55 +40,6 @@
 <%# Section for static/standalone pages (e.g. T&Cs) %>
 <%= render partial: 'partials/standalone' %>
 
-
-<%# statically written menu template used for each item menu %>
-<ul class="component-activated-menu govuk-navigation"
-    data-activator-classname="flow-add-page-button"
-    data-activator-text="<%= t('pages.create') %>"
-    data-component="PageAdditionMenu">
-  <li>
-    <span>Single question page</span>
-    <ul class="govuk-navigation">
-      <li data-page-type="singlequestion"
-          data-component-type="text"><a href="#add-page">Text</a></li>
-      <li data-page-type="singlequestion"
-          data-component-type="textarea"><a href="#add-page">Textarea</a></li>
-      <li data-page-type="singlequestion"
-          data-component-type="email"><a href="#add-page">Email address</a></li>
-      <li data-page-type="singlequestion"
-          data-component-type="number"><a href="#add-page">Number</a></li>
-      <li data-page-type="singlequestion"
-          data-component-type="date"><a href="#add-page">Date</a></li>
-      <li data-page-type="singlequestion"
-          data-component-type="radios"><a href="#add-page">Radio buttons</a></li>
-      <li data-page-type="singlequestion"
-          data-component-type="checkboxes"><a href="#add-page">Checkboxes</a></li>
-      <li data-page-type="singlequestion"
-          data-component-type="upload"><a href="#add-page">File upload</a></li>
-    </ul>
-  </li>
-  <li data-page-type="multiplequestions"><a href="#add-page">Multiple question page</a></li>
-  <% if ENV['BRANCHING'] == 'enabled' %>
-    <% unless service.checkanswers_page.present? %>
-      <li data-page-type="checkanswers"><a href="#add-page">Check answers page</a></li>
-    <% end %>
-    <% unless service.confirmation_page.present? %>
-      <li data-page-type="confirmation"><a href="#add-page">Confirmation page</a></li>
-    <% end %>
-  <% else %>
-    <% unless service.checkanswers_page.present? %>
-      <li data-page-type="checkanswers"><a href="#add-page">Check answers page</a></li>
-    <% end %>
-    <% unless service.confirmation_page.present? %>
-      <li data-page-type="confirmation"><a href="#add-page">Confirmation page</a></li>
-    <% end %>
-  <% end%>
-  <li data-page-type="content"><a href="#add-page">Content page</a></li>
-  <% if ENV['BRANCHING'] == 'enabled' %>
-    <li data-page-type="exit"><a href="#add-page">Exit page</a></li>
-  <% end %>
-</ul>
-
 <%# Add page component dialog template %>
 <div class="component-dialog-form"
      data-component="PageAdditionDialog"
